Replace deprecated Buffer constructor
The deprecated Buffer constructor has been replaced with `Buffer.from`
and `Buffer.alloc`. This removes deprecation warnings when this library
is used.

There were some instances where either the Buffer constructor or
`Buffer.from` was used, depending on whether `Buffer.from` was defined.
But this was unnecessary because the `engines` field of this package
ensures that the minimum Node.js version used is 6, which includes the
`Buffer.from` method.